requirements.txt
# FitLife AI 2.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
google-generativeai>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
